Crate eviov

The main common crate for the client and servers.

This crate reexports items defined in other subcrates.

Re-exports

pub use eviov_proto as proto;

pub use eviov_transport as transport;

Modules

orbit

tokio
Implementation for server crates compiling to native.

wasm
Implementation for client crates compiling to WASM.

Structs

Context
User wrapper for ContextImpl.

Eci
Denotes the ECI position and velocity of an object.

Length
Represents a length value.

Mass
Represents a mass value.

ObjectId
A network-unique runtime identifier for objects.

Orbit
Denotes an orbit, sensitive to initial time (the true anomaly).

OrbitInstant
The information calculated about the orbit at the specified instant.

ScaleVector
A special type of vector dedicated for scaling.

Time
The coordinated time in the game universe.

Vector
A general-purpose 2D vector.

Enums

LoopAction
General-purpose enum to denote the termination action of a looping function.

Constants

MILLIS_PER_TICK
The number of milliseconds per tick.

Traits

ContextImpl
Context-dependent primitives.