Module eviov::proto

Internal connection protocols.

There are 5 types of connections, as in the submodules time, cs, intra, ch and sh. Refer to the documentation of those submodules for details.

This module also exposes a number of top-level traits, to be implemented by protocol implementations. These traits are intended to be implemented by the codegen::proto! macro; do not try to implement them by hand.

Modules

ch

cs_ctrl
The control channel

cs_obs
The observer channel

intra

sh

time

transport

Structs

QueryId
Wraps the ID of a query.

Traits

ClientEndpoint
A marker trait indicating that the endpoint is the client side of the protocol

Endpoint
An enum covering all possible messages from a protocol endpoint.

Handler
A generic handler trait for protocols.

Message
Represents a message that can be sent through the protocol.

MessageFrom
A marker trait indicating that the message can be sent through a particular endpoint.

Protocol
A type representing a connection protocol.

QueryRequest
A message that expects a response value.

QueryRequestFrom
A marker trait indicating that the message can be sent through a particular endpoint.

QueryResponse
A message that responds to a QueryRequest message.

ServerEndpoint
A marker trait indicating that the endpoint is the server side of the protocol

Single
A marker trait indicating that the message is single, contrary to a request-response query pair.