Module eviov::proto::cs

Modules

ctrl
The control channel

obs
The observer channel

Structs

ChannelId

CtrlClient
Wraps a message on an ctrl channel from client to server

CtrlServer
Wraps a message on an ctrl channel from server to client

ObsClient
Wraps a message on an obs channel from client to server

ObsServer
Wraps a message on an obs channel from server to client

OpenChannelRequest
Ooens a new channel with the specified system

OpenChannelResponse

Proto

Enums

ChannelType

FromClient

FromServer