Module eviov::proto::cs_obs

The observer channel

Structs

ClientHandlerWrapper

Close
Acknowledges thst the channnel is closed, especially if the client is denied from observation.

Event
Reports a change of object states in the system.

HandshakeRequest
Identifies the client to acknowledge that it is authorized to observe

HandshakeResponse

Proto

ServerHandlerWrapper

Enums

Client

EventContent

Server

Traits

ClientHandler

ServerHandler