Trait eviov::proto::Endpoint

pub trait Endpoint: 'static + Send + Sync + Debug + Deserialize<'de> + Serialize where
    <Self::Peer as Endpoint>::Protocol == Self::Protocol,
    <Self::Peer as Endpoint>::Peer == Self, {
    type Protocol: Protocol;
    type Peer: Endpoint;
    fn request_query_id(&self) -> Option<QueryId>;

    fn response_query_id(&self) -> Option<QueryId>;
}

An enum covering all possible messages from a protocol endpoint.

Associated Types

type Protocol: Protocol

The protocol that this endpoint belongs to.

type Peer: Endpoint

The othe side of the protocol, i.e. Client => Server, Server => Client

Loading content...
Required methods

fn request_query_id(&self) -> Option<QueryId>

Retrieves the request query ID of the enum value.

If the enum value represents a query request, this returns the query ID of the request. Otherwise, this returns None.

fn response_query_id(&self) -> Option<QueryId>

Retrieves the response query ID of the enum value.

If the enum value represents a query response, this returns the query ID of the response. Otherwise, this returns None.

Loading content...