Trait eviov::proto::MessageFrom

pub trait MessageFrom<E: Endpoint>: Message<Protocol = <E as Endpoint>::Protocol> {
    fn from_enum(e: E) -> Option<Self>;

    fn to_enum(self) -> E;
}

A marker trait indicating that the message can be sent through a particular endpoint.

Required methods

fn from_enum(e: E) -> Option<Self>

Converts an enum of message from the endpoint to this particular type, returning None if the enum value is incorrect.

fn to_enum(self) -> E

Converts this message to its enum representation for the endpoint.

Loading content...