Trait eviov::proto::QueryRequest

pub trait QueryRequest: Message {
    fn query_id(&self) -> QueryId;

    fn set_query_id(&mut self, id: QueryId);
}

A message that expects a response value.

Implementors of this trait must also implement QueryRequestFrom<E> for at least one endpoint E; if a request type can be sent from both endpoints, it should implement QueryRequestFrom<E> twice, each E corresponding to each endpoint.

Required methods

fn query_id(&self) -> QueryId

The query ID of this message.

This method is internal, and should only be called by the connection handler in the transport submodule.

fn set_query_id(&mut self, id: QueryId)

Sets the query ID of this message.

This method is internal, and should only be called by the connection handler in the transport submodule.

Loading content...